Create PassingFreeParkingEvent.java
Create PassingFreeParkingEvent class.

A passing free parking event happens if the user is currently guided by the Parking Guidance System (PGS) and a free parking space is passed (due to lack of sensor at this parking).

Note: the event is currently always thrown and does not care whether the parking space is interesting for the user or not.